Vending Machine Program in Java
I have designed this program as an Assignment given in our college...
The Problem statement looks like this-
Consider a customer visiting a Vending Machine to buy some products. You can define your own products and their prices along with the denominations available to pay(for example, 50 rupee notes, 25 rupee notes, etc.). Calculate the total expense and input the amount that the customer pays. Design a vending machine which displays the change and optimal number of notes/coins to be paid as change.

(I have taken 10 & 20 rupees notes as an Example in this program. You can change their values.)

Star and Fork.

Thankyou.